Improve seeds, configuration and events endpoints
