+++ .iex.exs
@@ -9,6 +9,9 @@ alias Bokken.{
   Accounts.Ninja,
   Accounts.Organizer,
   Accounts.User,
+  Documents,
+  Documents.File,
+  Events,
   Events.Event,
   Events.Events,
   Events.Lecture,
