+++ .tool-versions
@@ -1,2 +1,2 @@
-erlang 24.0.5
-elixir 1.12.2-otp-24
+erlang 24.0.6
+elixir 1.12.3-otp-24
