+++ lib/bokken/mailer.ex
@@ -2,5 +2,5 @@ defmodule Bokken.Mailer do
   @moduledoc """
   A mailer module providing functionality to deliver emails.
   """
-  use Bamboo.Mailer, otp_app: :bokken
+  use Swoosh.Mailer, otp_app: :bokken
 end
