+++ lib/bokken_web/auth/pipeline.ex
@@ -13,5 +13,6 @@ defmodule BokkenWeb.Auth.Pipeline do
   plug Guardian.Plug.VerifyHeader
   plug Guardian.Plug.EnsureAuthenticated
   plug Guardian.Plug.LoadResource
+
   plug BokkenWeb.Auth.CurrentUser
 end
