+++ lib/bokken_web/controllers/mentor_controller.ex
@@ -7,7 +7,9 @@ defmodule BokkenWeb.MentorController do
 
   action_fallback BokkenWeb.FallbackController
 
-  def index(conn, params) do
+  def index(conn, params)
+      when is_map_key(params, "team_id")
+      when is_map_key(params, "event_id") do
     mentors = Accounts.list_mentors(params)
     render(conn, "index.json", mentors: mentors)
   end
