+++ lib/bokken_web/templates/email/verify.text.eex
@@ -5,7 +5,7 @@ que te aparece a seguir.
 
 Copia este link para o teu browser:
 
-                          <%= @link %>
+<%= @link %>
 
 Estamos contentes por teres demonstrado interesse em fazer parte do CoderDojo
 Braga. Estamos ansiosos por te ver na próxima sessão.
